python: move np_array_to_list to utils, fix CMakeLists
