+++ interfaces/CMakeLists.txt
@@ -59,7 +59,7 @@ add_test(NAME python_test_ocp
         python main_test.py)
 add_test(NAME python_pendulum_example
         COMMAND "${CMAKE_COMMAND}" -E chdir ${PROJECT_SOURCE_DIR}/examples/acados_python/pendulum_example
-        python minimal_example.py)
+        python minimal_example_ocp.py)
 add_test(NAME python_soft_constraints_example
         COMMAND "${CMAKE_COMMAND}" -E chdir ${PROJECT_SOURCE_DIR}/examples/acados_python/soft_constraints_example
         python generate_c_code.py)
