+++ interfaces/acados_template/acados_template/acados_ocp_nlp.py
@@ -33,7 +33,6 @@
 #
 
 import numpy as np
-import casadi as ca
 import json
 import os
 import sys
@@ -1894,19 +1893,6 @@ class acados_ocp_nlp:
         setter_to_call(tokens[1], value)
 
         return
-
-
-def np_array_to_list(np_array):
-    if  isinstance(np_array, (np.ndarray)):
-        return np_array.tolist()
-    elif  isinstance(np_array, (ca.SX)):
-        return ca.DM(np_array).full()
-    elif  isinstance(np_array, (ca.DX)):
-        return np_array.full()
-    else:
-        raise(Exception(
-            "Cannot convert to list type {}".format(type(np_array))
-        ))
 
 class ocp_nlp_as_object:
         def __init__(self, d):
